+++ src/full/Agda/TypeChecking/Serialise.hs
@@ -82,7 +82,7 @@ import Agda.Utils.Impossible
 -- 32-bit machines). Word64 does not have these problems.
 
 currentInterfaceVersion :: Word64
-currentInterfaceVersion = 20221024 * 10 + 0
+currentInterfaceVersion = 20221027 * 10 + 0
 
 -- | The result of 'encode' and 'encodeInterface'.