+++ CHANGELOG.md
@@ -142,6 +142,12 @@ Cubical Agda
 * Rewrite rules with interval arguments are now supported.
   ([#4384](https://github.com/agda/agda/issues/4384))
 
+* Matching on `@flat` arguments is now disabled by default when the
+  `--cubical-compatible` or `--without-K` flags are used. It can be
+  enabled again using the `--flat-split` flag. Note that, when using
+  Cubical Agda, functions which match on a `@flat` argument will not
+  compute when applied to `transport`s.
+
 Reflection
 ----------